Updated and norminette everything
