+++ src/pipex/execute_2.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   execute_2.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gyong-si <gyong-si@student.42.fr>          +#+  +:+       +#+        */
+/*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/26 19:30:03 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/28 23:19:59 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/06/29 11:32:21 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,6 @@ void	execute_pipeline(t_token *curr, t_shell *minishell)
 	int	pipe_fd[2];
 	int	pid;
 
-	//printf("execute_pipeline\n");
 	if (pipe(pipe_fd) == -1)
 		exit(EXIT_FAILURE);
 	pid = fork();
